requirements.txt
django==2.2.*
djangorestframework==3.*
djoser==2.2.1